Hard fork of the Phazerville tree onto the one board it runs on, plus the
instrument work the exploration said was needed and the bench then argued
with.

Measure first (track A): a written timing budget with eleven ceilings and
the counters to check them, reported by a new console 'T'. Three ISR-context
bugs fixed on the way: clock Start/Stop sent USB MIDI from the CORE ISR, an
audio-CPU overload switched audio off for a quarter second, and the
ISR-to-loop defer path allocated inside the ISR.

Delete second (track B): 23 vestigial build environments, the 1776-line
2-up host, four apps, and every preprocessor branch for hardware this fork
does not run on. Six grep gates in CI keep them from coming back.

Then the instrument (track C): recall no longer writes to flash at all
(630 ms to 2 ms, served from a RAM stage and synced at idle), saves declare
a persistence window and fade the codec across it, the bus MIDI ring has a
realtime lane, the 200e card image writes only the sectors that changed,
and a broadcast RECALL decodes which modules answered.

Two planned items were RETIRED by measurement rather than built: the
load-shedding governor (the CORE ISR peaks at 31 us of its 60 us period, so
there is nothing to shed) and the removal of the per-tick audio interrupt
mask (0 dropped blocks and a worst ISR gap of exactly one period, including
under a live parameter sweep). Both are written up in docs/Timing-Budget.md
under What the budget has retired. That was the point of writing the
ceilings down first: to decide with a number instead of an intuition, and
both intuitions were wrong.

The bench found what the host tests and the simulator could not. The window
timed itself with millis(), which is masked during the very write it
measures. AUDIO_INTERFACE is a USB descriptor number and was being read as
"this build has audio" in 30 places, which had compiled the save fade out of
the console build, left the Tuner deaf, and left the panel audio outputs
unbuilt at boot. And the Tuner leaked about 11 audio blocks per preset save
until the pool ran dry and the codec went permanently silent, which nothing
short of a counter would have caught.

Verified on hardware (T41_console): 0 of 11 budget rows fail at idle and
across saves; save 127-196 ms inside its declared 250 ms window; recall
2 ms; the 200e bus queried, a RECALL broadcast and a 259e bank backed up
without disturbing the master; no crashes across many flashes.

All five environments build, 25 host suites pass, the simulator self-check
is green, and all six fork gates hold.

+++ software/src/apps/ScaleEditor.h
@@ -162,7 +162,16 @@ private:
     int8_t cursor;
     int8_t current_scale;
     int8_t current_note;
-    int8_t current_import_scale;
+    // NOT int8_t: this indexes the whole scale table, 0..NUM_SCALES-1, and
+    // NUM_SCALES is 148 (4 user slots + 144 braids scales) -- past what an
+    // int8_t holds. The import picker constrains to NUM_SCALES-1 = 147 and
+    // then assigns that back here, so stepping up from 127 stored -128, which
+    // (a) put the top 20 scales permanently out of reach, and (b) indexed
+    // scale_names[-128] on the very next draw -- an out-of-bounds read on the
+    // display path, not just a stuck cursor. GetScale() re-CONSTRAINs and so
+    // never saw it, which is why this stayed quiet.
+    // Transient UI state, not serialised, so widening it changes no format.
+    int16_t current_import_scale;
     int undo_value;
     bool length_set_mode;
     bool import_mode;
